requirements.txt
python-dotenv
elevenlabs
langsmith
langchain
langchain-openai
openai
pydantic
streamlit
pandas
fastapi[standard]
uvicorn
python-multipart
minio
pymongo
langchain_google_genai